+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Footer from "./components/Footer.tsx";
-import Home from "./components/home/Home.tsx";
+import Home from "./components/Home.tsx";
 import PageNotFound from './components/PageNotFound.tsx';
 import { Navbar } from "./components/nav/navbar";
 import SaiBaba from "./components/sai-baba.tsx";
