+++ src/components/Footer.tsx
@@ -30,6 +30,7 @@ function Footer() {
             <ul>
               <li className="mb-2"><a href="/" className="text-gray-200 hover:text-white">Home</a></li>
               <li className="mb-2"><a href="https://www.saiparivarfoundation.org/about.html" target={"_blank"}>Sai Baba</a></li>
+              <li className="mb-2"><a href="https://www.saiparivarfoundation.org/guruji.html" target={"_blank"}>Guruji</a></li>
               <li className="mb-2"><a href="https://www.saiparivarfoundation.org/mission.html" target={"_blank"}>About us</a></li>
               <li className="mb-2"><a href="https://www.paypal.com/paypalme/SaiParivarFoundation"  target={"_blank"} className="text-gray-200 hover:text-white">Donate</a></li>
             </ul>
