+++ src/components/Home.tsx
@@ -2,8 +2,9 @@ import React from "react";
 
 import Section1 from "./Section1.tsx";
 import Section2 from "./Section2.tsx";
-import SaiBaba from "../sai-baba.tsx";
-import AboutUs from "../AboutUs.tsx";
+import SaiBaba from "./sai-baba.tsx";
+import AboutUs from "./AboutUs.tsx";
+import Guruji from "./guruji.tsx";
 
 function Home() {
   return (
@@ -11,6 +12,7 @@ function Home() {
       <Section1 />
       <Section2 />
       <SaiBaba />
+        <Guruji />
       <AboutUs />
     </>
   );
