+++ src/components/nav/navbar/mobileNavLinks.jsx
@@ -56,7 +56,7 @@ export function MobileNavLinks() {
                 <Link href="https://www.saiparivarfoundation.org/about.html" target={"_blank"}>Sai Baba</Link>
             </LinkItem>
             <LinkItem>
-                <Link href="/guruji">Guruji</Link>
+                <Link href="https://www.saiparivarfoundation.org/guruji.html" target={"_blank"}>Guruji</Link>
             </LinkItem>
             <LinkItem>
                 <Link href="https://www.saiparivarfoundation.org/mission.html" target={"_blank"}>About us</Link>
